Website+Camera/robots-main/picode/requirements.txt
pyserial
paho-mqtt
opencv-python
numpy
